backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
sqlalchemy
pydantic
Office365-REST-Python-Client
python-multipart
python-dotenv
xhtml2pdf
